requirement.txt
fpdf2   >=  2.5.7
Pillow  >=  9.0.0
PyPDF2  >=  2.11.1